# servants-of-india/docs/openapi.swagger.yaml
...
openapi: 3.0.3
info:
  title: Servants of Bharat API — Complete (All APIs)
  description: Swagger / OpenAPI 3.0 documentation for ALL APIs created for the Servants of Bharat platform.
    Every operation includes its mapped user story (x-user-story) and error-handling details (x-error-handling).
  version: 1.0.0
servers:
- url: /api
  description: Local API base path
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Forbidden: insufficient role'
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        organization:
          type: string
          nullable: true
        role:
          type: string
          enum:
          - volunteer
          - event_manager
          - super_admin
        status:
          type: string
          enum:
          - active
          - blocked
          - deactivated
        created_at:
          type: string
          format: date-time
...
paths:
...
  /users/me:
    get:
      tags:
      - Users
      summary: Get current profile
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '401': &id001
          description: Missing/invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403': &id002
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
      x-user-story: As any user, I want to view my profile so that I can see my current details.
      x-error-handling:
      - 401 - missing or invalid JWT
      description: 'Get current profile


        **User story:** As any user, I want to view my profile so that I can see my current details.


        **Error handling:**

        - 401 - missing or invalid JWT'
    put:
      tags:
      - Users
      summary: Update own profile (email immutable)
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '401': *id001
        '403': *id002
      security:
      - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      x-user-story: As any user, I want to update my profile so that my details stay current.
      x-error-handling:
      - 401 - not authenticated
      - 422 - full_name cannot be empty
      description: 'Update own profile (email immutable)


        **User story:** As any user, I want to update my profile so that my details stay current.


        **Error handling:**

        - 401 - not authenticated

        - 422 - full_name cannot be empty'
  /users/me/password:
    put:
      tags:
      - Users
      summary: Change own password (stays logged in)
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: New password too short
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': *id001
        '403': *id002
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - current_password
              - new_password
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      x-user-story: As any user, I want to change my password securely so that my account stays safe.
      x-error-handling:
      - 400 - current password incorrect (does NOT log the user out)
      - 422 - new password shorter than 8 characters
      - 401 - not authenticated
      description: 'Change own password (stays logged in)


        **User story:** As any user, I want to change my password securely so that my account stays safe.


        **Error handling:**

        - 400 - current password incorrect (does NOT log the user out)

        - 422 - new password shorter than 8 characters

        - 401 - not authenticated'
  /users:
    get:
      tags:
      - Users
      summary: List all users (Super Admin)
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '401': *id001
        '403': *id002
      security:
      - bearerAuth: []
      parameters:
      - name: role
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
      x-user-story: As a super admin, I want to list and filter all users so that I can administer accounts.
      x-error-handling:
      - 401 - not authenticated
      - 403 - caller is not a Super Admin
      description: 'List all users (Super Admin)


        **User story:** As a super admin, I want to list and filter all users so that I can administer
        accounts.


        **Error handling:**

        - 401 - not authenticated

        - 403 - caller is not a Super Admin'
    post:
      tags:
      - Users
      summary: Create Event Manager / Super Admin (Super Admin)
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': *id001
        '403': *id002
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - full_name
              - email
              - password
              - role
              properties:
                full_name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum:
                  - event_manager
                  - super_admin
                phone:
                  type: string
                location:
                  type: string
                organization:
                  type: string
      x-user-story: As a super admin, I want to create Event Manager / Super Admin accounts so that the
        platform can be staffed.
      x-error-handling:
      - 422 - validation error (EM requires organization, phone, location)
      - 409 - email already registered
      - 403 - caller is not a Super Admin
      description: 'Create Event Manager / Super Admin (Super Admin)


        **User story:** As a super admin, I want to create Event Manager / Super Admin accounts so that
        the platform can be staffed.


        **Error handling:**

        - 422 - validation error (EM requires organization, phone, location)

        - 409 - email already registered

        - 403 - caller is not a Super Admin'
  /users/{id}/status:
    patch:
      tags:
      - Users
      summary: Change user status (Super Admin)
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '401': *id001
        '403': *id002
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - status
              properties:
                status:
                  type: string
                  enum:
                  - active
                  - blocked
                  - deactivated
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      x-user-story: As a super admin, I want to block / unblock / deactivate users so that I control access.
      x-error-handling:
      - 400 - cannot change your own status
      - 422 - invalid status value
      - 404 - user not found
      - 403 - caller is not a Super Admin
      description: 'Change user status (Super Admin)


        **User story:** As a super admin, I want to block / unblock / deactivate users so that I control
        access.


        **Error handling:**

        - 400 - cannot change your own status

        - 422 - invalid status value

        - 404 - user not found

        - 403 - caller is not a Super Admin'
...